+++ web/src/components/ui/index.ts
@@ -588,3 +588,4 @@ export * from './chart-line-vroc-cross-sig';
 export * from './chart-line-macd-cross-sig';
 export * from './chart-line-stoch-cross-pct';
 export * from './chart-line-rsi-cross-pct';
+export * from './chart-line-mfi-cross-pct';